Refactor: split into separate examples
