+++ go/adbc/driver/salesforce/gosalesforce/shared/utils.go
